ref: move AuthorNotingHook in tp-traits
