+++ primitives/traits/src/lib.rs
@@ -19,9 +19,36 @@
 
 #![cfg_attr(not(feature = "std"), no_std)]
 
-pub use cumulus_primitives_core::{relay_chain::Slot, ParaId};
-use frame_support::pallet_prelude::DispatchResultWithPostInfo;
+pub use cumulus_primitives_core::{
+    relay_chain::{BlockNumber, Slot},
+    ParaId,
+};
+use frame_support::pallet_prelude::{DispatchResultWithPostInfo, Weight};
 use sp_std::vec::Vec;
+
+/// The author-noting hook to react to container chains authoring.
+pub trait AuthorNotingHook<AccountId> {
+    /// This hook is called partway through the `set_latest_author_data` inherent in author-noting.
+    ///
+    /// The hook should never panic and is required to return the weight consumed.
+    fn on_container_author_noted(
+        author: &AccountId,
+        block_number: BlockNumber,
+        para_id: ParaId,
+        relay_block_number: BlockNumber,
+    ) -> Weight;
+}
+
+impl<AccountId> AuthorNotingHook<AccountId> for () {
+    fn on_container_author_noted(
+        _: &AccountId,
+        _: BlockNumber,
+        _: ParaId,
+        _: BlockNumber,
+    ) -> Weight {
+        Weight::zero()
+    }
+}
 
 pub trait DistributeRewards<AccountId, Balance> {
     fn distribute_rewards(rewarded: AccountId, amount: Balance) -> DispatchResultWithPostInfo;
